# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Noted API — collaborative human-made music platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /songs/{songId}/comments:
    get:
      operationId: listSongComments
      tags: [comments]
      summary: List comments on a song, newest first
      parameters:
        - $ref: "#/components/parameters/SongId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Comment" }
    post:
      operationId: postSongComment
      tags: [comments]
      summary: Post a comment on a song (auth required)
      parameters:
        - $ref: "#/components/parameters/SongId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostCommentBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (author or admin)
      parameters:
        - name: commentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
  /comments/{commentId}/reports:
    post:
      operationId: reportComment
      tags: [comments]
      summary: Flag a comment for moderation (auth required, idempotent per user)
      parameters:
        - name: commentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReportCommentBody" }
      responses:
        "201":
          description: Reported
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
...
components:
  parameters:
    SongId:
      in: path
      name: songId
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    ContributorLite:
      type: object
      required: [id, displayName]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        username:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        socialHandle:
          type: string
          nullable: true
...
    Comment:
      type: object
      required: [id, songId, authorId, body, createdAt, author]
      properties:
        id: { type: string, format: uuid }
        songId: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        body: { type: string }
        createdAt: { type: string, format: date-time }
        author: { $ref: "#/components/schemas/ContributorLite" }
    PostCommentBody:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 2000 }
    ReportCommentBody:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 1, maxLength: 500 }
...